refactor structure of `test` directory
