+++ packages/patrol_cli/test/commands/patrol_command_runner_test.dart
@@ -8,7 +8,7 @@ import 'package:patrol_cli/src/runner/patrol_command_runner.dart';
 import 'package:pub_updater/pub_updater.dart';
 import 'package:test/test.dart';
 
-import 'mocks.dart';
+import '../src/mocks.dart';
 
 const latestVersion = '0.0.0';
