+++ packages/patrol_cli/test/general/pubspec_reader_test.dart
