+++ packages/patrol_cli/test/src/fakes.dart
