+++ packages/patrol_cli/test/src/mocks.dart
